requirements.txt
numpy
oauth2client
streamlit==1.32.2
gspread==6.1.1
gspread-dataframe  
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
bcrypt==4.1.3
pandas 
xlwt==1.3.0  
openpyxl  
pillow  
pytz
python-dateutil==2.8.2
python-dateutil
tkcalendar
pydrive2
streamlit_autorefresh